% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{armor}
\alias{armor}
\title{Armor information}
\format{A data.frame with a whole bunch of variables}
\usage{
armor
}
\description{
All of the information about DND 5e armor
}
\keyword{datasets}

