% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d20.R
\name{d20}
\alias{d20}
\alias{d4}
\alias{d6}
\alias{d8}
\alias{d10}
\alias{d12}
\alias{d100}
\title{Roll dice}
\usage{
d20(n)

d4(n)

d6(n)

d8(n)

d10(n)

d12(n)

d100(n)
}
\arguments{
\item{n}{A scalar integer indicating the number of dice to roll}
}
\value{
A vector of dice rolls
}
\description{
Roll some dice
}
\section{Functions}{
\itemize{
\item \code{d4()}: Roll n d4 dice

\item \code{d6()}: Roll n d6 dice

\item \code{d8()}: Roll n d8 dice

\item \code{d10()}: Roll n d10 dice

\item \code{d12()}: Roll n d12 dice

\item \code{d100()}: Roll n d100 dice

}}
\examples{
d20(n=2)

}
