% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_monsters.R
\name{get_monsters}
\alias{get_monsters}
\title{Get certain monsters}
\usage{
get_monsters(name = NULL)
}
\arguments{
\item{name}{a character vector of monster names}
}
\value{
data.frame of monsters
}
\description{
Get monsters by name
}
\examples{
get_monsters(name = "orc")

}
