% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spells.R
\name{get_spells}
\alias{get_spells}
\title{Get certain spells}
\usage{
get_spells(class = NULL, level = NULL, school = NULL)
}
\arguments{
\item{class}{a character vector of classes}

\item{level}{an integer vector of levels}

\item{school}{a character vector of schools}
}
\value{
data.frame of spells
}
\description{
Get spells by classes, levels, schools, etc
}
\examples{
get_spells(class = c("druid"), level = 1)

}

