% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hp.R
\name{hp}
\alias{hp}
\title{Roll hit points}
\usage{
hp(hp, n = 1)
}
\arguments{
\item{hp}{character AdB+C}

\item{n}{number of creatures to}
}
\value{
numeric vector of hit points
}
\description{
Roll hit points based on the standard formula AdB+C where you roll A hit 
dice with B sides and add C
}
