% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{melee_weapons}
\alias{melee_weapons}
\title{Melee weapon details}
\format{A data.frame with a whole bunch of variables}
\usage{
melee_weapons
}
\description{
All of the information about DND 5e melee weapons
}
\keyword{datasets}

