% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{monsters}
\alias{monsters}
\title{Monsters details}
\format{A data.frame with a whole bunch of variables}
\usage{
monsters
}
\description{
All of the information about DND 5e monsters.
}
\keyword{datasets}

