% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ranged_weapons}
\alias{ranged_weapons}
\title{Ranged weapon details}
\format{
A data.frame with a whole bunch of variables
}
\usage{
ranged_weapons
}
\description{
All of the information about DND 5e ranged weapons
}
\keyword{datasets}
