% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll_ability_scores.R
\name{roll_ability_scores}
\alias{roll_ability_scores}
\title{Roll for ability scores}
\usage{
roll_ability_scores(n = 6)
}
\arguments{
\item{n}{The number ability scores to roll (default=6).}
}
\value{
A numeric vector of ability scores
}
\description{
Roll for ability scores
}
\examples{
roll_ability_scores()
hist(roll_ability_scores(1e5), seq(2.5,20.5,by=1))

}
