% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll_attack.R
\name{roll_attack}
\alias{roll_attack}
\title{Roll attack}
\usage{
roll_attack(to_hit, to_damage, type = "normal", ...)
}
\arguments{
\item{to_hit}{amount to add to the d20 attack roll}

\item{to_damage}{amount to add to the damage roll}

\item{type}{normal, advantage, or disadvantage}

\item{...}{dice to roll, passed to \code{\link{roll_dice}}}
}
\value{
attack and damage roll are printed
}
\description{
Roll an attack
}
