% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll_dice.R
\name{roll_dice}
\alias{roll_dice}
\title{Roll dice}
\usage{
roll_dice(d4 = NULL, d6 = NULL, d8 = NULL, d10 = NULL, d20 = NULL,
  d100 = NULL)
}
\arguments{
\item{d4}{A positive integer indicating the number of d4 to roll}

\item{d6}{A positive integer indicating the number of d6 to roll}

\item{d8}{A positive integer indicating the number of d8 to roll}

\item{d10}{A positive integer indicating the number of d10 to roll}

\item{d100}{A positive integer indicating the number of d100 to roll}

\item{d12}{A positive integer indicating the number of d20 to roll}
}
\description{
Allows you to roll any combination of dice
}
\examples{
roll_dice(d4=2)
roll_dice(d4=2,d6=3)

}

