% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll_hp.R
\name{roll_hp}
\alias{roll_hp}
\title{Roll hit points}
\usage{
roll_hp(hp, n = 1)
}
\arguments{
\item{hp}{character AdB+C}

\item{n}{number of creatures to}
}
\value{
numeric vector of hit points
}
\description{
Roll hit points based on the standard formula AdB+C where you roll A hit 
dice with B sides and add C
}
\examples{
roll_hp("2d4+1")
roll_hp("4d10+12", 10)


}
