% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll_npc.R
\name{roll_npc}
\alias{roll_npc}
\title{Roll NPC}
\usage{
roll_npc()
}
\value{
a data.frame containing details about an NPC including 
appearance, abilities, talent, mannerism, interaction trait, ideal,
bond, and flaw or secret.
}
\description{
Create a full-featured NPC
}
