% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll_npc.R
\name{roll_npc_abilities}
\alias{roll_npc_abilities}
\title{NPC Abilities}
\usage{
roll_npc_abilities()
}
\value{
a data.frame containing the high and low abilities
}
\description{
Random high and low abilities
}
