% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll_npc.R
\name{roll_npc_flaw_or_secret}
\alias{roll_npc_flaw_or_secret}
\title{NPC flaw or secret}
\usage{
roll_npc_flaw_or_secret()
}
\value{
a data.frame with a random flaw or secret
}
\description{
Random NPC flaw or secret
}
