% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll_npc.R
\name{roll_npc_interaction_trait}
\alias{roll_npc_interaction_trait}
\title{NPC Interaction trait}
\usage{
roll_npc_interaction_trait()
}
\value{
a data.frame with an interaction trait
}
\description{
Random npc interaction trait
}
