% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spell_info}
\alias{spell_info}
\title{Spell information}
\format{A data.frame with a whole bunch of variables}
\usage{
spell_info
}
\description{
All of the information about DND 5e spells
}
\keyword{datasets}

