% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spells}
\alias{spells}
\title{Spell information}
\format{
A data.frame with a whole bunch of variables
}
\usage{
spells
}
\description{
All of the information about DND 5e spells
}
\keyword{datasets}
