% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treasure.R
\name{treasure}
\alias{treasure}
\title{Roll treasure}
\usage{
treasure(n, challenge = 0, type = "individual", collapse = FALSE)
}
\arguments{
\item{n}{number of times to roll}

\item{challenge}{challenge rating of creatures}

\item{type}{individual or hoard}

\item{collapse}{logicical, if true, will just return total}
}
\description{
Roll for random treasure
}
\examples{
treasure(n = 2) # 2 individuals challenge of 0
treasure(n = 1, challenge = 3)
treasure(n=29, collapse = TRUE)

}
